Export some env vars, proxy D-Bus session bus
Export all current env vars (except $HOME as that could contain /root),
but it could possibly include $DISPLAY and $DBUS_SESSION_BUS_ADDRESS.

DBUS_SESSION_BUS_ADDRESS is set to a named Unix socket address when an
abstract socket is found. This ensures that the net namespace will also
be able to connect to the D-Bus session bus.